xxm_fans_home/requirements.txt
acme==2.9.0
appdirs==1.4.4
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Bottleneck==1.3.5
Brotli==1.1.0
certbot==2.9.0
certbot-nginx==2.9.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==23.2.2
colorama==0.4.6
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.8
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
et-xmlfile==1.0.1
fonttools==4.46.0
fs==2.4.16
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==0.0.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.2.1
lz4==4.0.2+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mpmath==0.0.0
netifaces==0.11.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4+dfsg
parsedatetime==2.6
pcp==5.0
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyICU==2.12
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-linux-procfs==0.6.3
python-magic==0.4.27
pytz==2024.1
pyudev==0.24.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sos==4.5.6
soupsieve==2.5
ssh-import-id==5.11
supervisor==4.2.5
sympy==1.12
systemd-python==235
tables==3.9.2
Twisted==24.3.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
wadllib==1.3.6
webencodings==0.5.1
wheel==0.42.0
xkit==0.0.0
zope.interface==6.1
